+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `outlined` default value for `radio item` component
+- Add flip option for icon on the trailing action of the text input (Orange-OpenSource/ouds-ios#1041)
 - Add flip option for button with icon in layout (Orange-OpenSource/ouds-ios#1124) 
 - Update `text input` to display the label on two lines and in the middle when no placeholder and one line at top otherwise  (Orange-OpenSource/ouds-ios#1086)